UltimateDestroyer relies on a ConnectionAdapter predicate to determine whether DELETE with LIMIT is supported. This prevents it from breaking when used with the SQLite3Adapter. Removed acts_as_paranoid support – do that in your app.

+++ lib/ultimate_destroyer.rb
@@ -1,33 +1,38 @@
+module ActiveRecord
+  module ConnectionAdapters
+    class AbstractAdapter
+      def supports_delete_with_limit?
+        false
+      end
+    end
+    
+    class MysqlAdapter
+      def supports_delete_with_limit?
+        true
+      end
+    end
+  end
+end
+
 module UltimateDestroyer
   module Model
     def self.included(base)
       base.extend ClassMethods
-      
       base.class_inheritable_accessor :destroy_amount
       base.destroy_amount = 500
-
-      class << base # this is fucking stupid
-        alias_method_chain :destroy_all, :no_mercy
-        alias_method_chain :delete_all, :no_mercy
-      end
     end
   
     module ClassMethods
-      
-    
-      def destroy_all_with_no_mercy(conditions = nil)
+      def destroy_all(conditions = nil)
+        return super unless connection.supports_delete_with_limit?
         until (records = find(:all, :conditions => conditions, :limit => destroy_amount)).blank?
           records.each { |record| record.destroy }
         end
       end
-
-      def destroy_all!(conditions = nil) # this is to work with acts_as_paranoid
-        until (records = find_with_deleted(:all, :conditions => conditions, :limit => destroy_amount)).blank?
-          records.each { |record| record.destroy! }
-        end
-      end
       
-      def delete_all_with_no_mercy(conditions = nil)
+      def delete_all(conditions = nil)
+        return super unless connection.supports_delete_with_limit?
+        
         sql = "DELETE FROM #{quoted_table_name} "
         add_conditions!(sql, conditions, scope(:find))
         sql += " LIMIT #{destroy_amount}"
@@ -39,29 +44,19 @@ module UltimateDestroyer
         end
         rows_deleted += just_deleted
       end
-      
     end
   end
   
   module Association
     def self.included(base)
       base.send(:include, InstanceMethods)
-      
-      base.send(:alias_method_chain, :destroy_all, :no_mercy)
     end
     
     module InstanceMethods
-      def destroy_all_with_no_mercy(conditions = nil)
+      def destroy_all(conditions = nil)
+        return super unless connection.supports_delete_with_limit?
         until (records = find(:all, :conditions => conditions, :limit => destroy_amount)).blank?
           records.each { |record| record.destroy }
-        end
-        
-        reset_target!
-      end
-
-      def destroy_all!(conditions = nil) # this is to work with acts_as_paranoid
-        until (records = find_with_deleted(:all, :conditions => conditions, :limit => destroy_amount)).blank?
-          records.each { |record| record.destroy! }
         end
         
         reset_target!
